Simplify: extract remaining backend api groups
